requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-environ==0.4.5
django-filebrowser==3.13.1
django-grappelli==2.14.2
django-js-asset==1.2.2
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.6
python3-openid==3.2.0
pytils==0.3
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
